+++ UserGrain/Models/UserState.cs
@@ -6,7 +6,7 @@ public class UserState
 
     public ulong CountChecked { get; set; }
     public ulong CountUnchecked { get; set; }
-    public int GoldBalance { get; set; }
+    public ulong GoldBalance { get; set; }
     public string? UserName { get; set; }
 
     #endregion
